removed async method and moved the source and test to project folder
