+++ .changeset/funny-socks-rescue.md
@@ -3,5 +3,5 @@
 '@sap-ux/fe-fpm-writer': patch
 ---
 
-Added validateId function to @sap-ux/project-access with sync/async overloads and reused it in @sap-ux/fe-fpm-writer.
-Exported findFilesByExtension from @sap-ux/project-access public API and removed deep import from @sap-ux/fe-fpm-writer.
+Added isUI5IdUnique function to check UI5 control ID uniqueness in XML views/fragments.
+Exported findFilesByExtension from @sap-ux/project-access and removed deep imports.
