+++ packages/fe-fpm-writer/src/common/file.ts
@@ -1,7 +1,7 @@
 import type { CopyOptions, Editor } from 'mem-fs-editor';
 import type { TabInfo } from '../common/types';
 import { sep, normalize } from 'node:path';
-import { findFilesByExtension, validateId } from '@sap-ux/project-access';
+import { findFilesByExtension, isUI5IdUnique } from '@sap-ux/project-access';
 
 /**
  * Options for creating an ID generator with cached file contents.
@@ -167,14 +167,16 @@ export const CONFIG = {
 function generateUniqueElementId(baseId: string, filteredFilesContent: string[], validatedIds: string[] = []): string {
     const maxAttempts = 1000;
 
-    if (validateId(baseId, validatedIds, { files: filteredFilesContent })) {
+    // Check both in-memory validatedIds and filesystem files
+    if (!validatedIds.includes(baseId) && isUI5IdUnique(baseId, filteredFilesContent)) {
         return baseId;
     }
 
     for (let counter = 1; counter < maxAttempts; counter++) {
         const candidateId = `${baseId}${counter}`;
 
-        if (validateId(candidateId, validatedIds, { files: filteredFilesContent })) {
+        // Check both in-memory validatedIds and filesystem files
+        if (!validatedIds.includes(candidateId) && isUI5IdUnique(candidateId, filteredFilesContent)) {
             return candidateId;
         }
     }
